bdev/raid: Check if required params are specified when parsing .INI config file
Checking if required parameters strip_size, num_base_bdevs, and
raid_level had been done in raid_bdev_config_add(). However, this
required us to use integer type.

If we do the check in raid_bdev_parse_raid(), we will be able to
use the same type between function parameters and variables of
struct, and hence do in this patch.

This will align with the next patch.

+++ lib/bdev/raid/bdev_raid.h
@@ -226,8 +226,8 @@ int raid_bdev_create(struct raid_bdev_config *raid_cfg);
 int raid_bdev_add_base_devices(struct raid_bdev_config *raid_cfg);
 void raid_bdev_remove_base_devices(struct raid_bdev_config *raid_cfg,
 				   raid_bdev_destruct_cb cb_fn, void *cb_ctx);
-int raid_bdev_config_add(const char *raid_name, int strip_size, int num_base_bdevs,
-			 int raid_level, struct raid_bdev_config **_raid_cfg);
+int raid_bdev_config_add(const char *raid_name, uint32_t strip_size, uint8_t num_base_bdevs,
+			 uint8_t raid_level, struct raid_bdev_config **_raid_cfg);
 int raid_bdev_config_add_base_bdev(struct raid_bdev_config *raid_cfg,
 				   const char *base_bdev_name, uint32_t slot);
 void raid_bdev_config_cleanup(struct raid_bdev_config *raid_cfg);
